semigrp: Add documentation and fix exceptions
Adds documentation for the Semigroup class and FullTransformationMonoid
function.

Adds exception to Semigroup __init__, when called with a Python
object, before passing to PythonElementNC, checks the element can be
multiplied by itself.

Adds error messages for exceptions.

Disables pylint len-as-condition, so that a list can be passed, as
well as a number of arguments not in a list.